Deprecate spa.SPA.
