+++ nengo/spa/tests/test_assoc_mem.py
@@ -16,7 +16,7 @@ def test_am_spa_interaction(Simulator, seed, rng):
     def input_func(t):
         return '0.49*A' if t < 0.5 else '0.79*A'
 
-    with nengo.spa.SPA(seed=seed) as m:
+    with nengo.spa.Module(seed=seed) as m:
         m.buf = nengo.spa.Buffer(D)
         m.input = nengo.spa.Input(buf=input_func)
